requirements.txt
flask==3.1.1
flask-cors==5.0.1
openai==1.97.0
python-dotenv==1.1.1
